backend/requirements.txt
# FastAPI Backend Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Core PitchOS Dependencies
google-generativeai>=0.3.0
PyPDF2>=3.0.1
python-docx>=0.8.11
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
pydantic>=2.0.0
typing-extensions>=4.7.0

# OCR Dependencies (Optional)
easyocr>=1.7.0
opencv-python>=4.8.0
google-cloud-vision>=3.4.0

# Additional FastAPI utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.0
